Rename InitAnt
