+++ src/Component/CommGS/InitAntenna.cpp
@@ -1,5 +1,5 @@
 #define _CRT_SECURE_NO_WARNINGS
-#include "InitAnt.hpp"
+#include "InitAntenna.hpp"
 
 #include <string.h>
